// src/types/dashboard/Reviewer/types.tsx
export type PaperAssesmentDataType = {
  topicComment: string;
  topic: number;
  contributionComment: string;
  contribution: number;
  academicQualityComment: string;
  academicQuality: number;
  verificationOfResultsComment: string;
  verificationOfResults: number;
  noveltyComment: string;
  novelty: number;
  literatureReviewAndBibliographyComment: string;
  literatureReviewAndBibliography: number;
  languageComment: string;
  language: number;
  styleAndFormatComment: string;
  styleAndFormat: number;
  summary: string;
  commentsForOrganizingCommittee: string;
  recommendation: string;
};